requirements.txt
PyPDF2>=3.0.0
python-docx>=1.1.0
pdfplumber>=0.10.0
PyYAML>=6.0
requests>=2.31.0
tqdm>=4.66.0
colorama>=0.4.6
